Added support for record count of more than 2 billions - by using long instead int.

+++ core/src/main/java/com/yahoo/ycsb/Client.java
@@ -237,7 +237,7 @@ class ClientThread extends Thread
   DB _db;
   boolean _dotransactions;
   Workload _workload;
-  int _opcount;
+  long _opcount;
   double _targetOpsPerMs;
 
   int _opsdone;
@@ -259,7 +259,7 @@ class ClientThread extends Thread
    * @param targetperthreadperms target number of operations per thread per ms
    * @param completeLatch The latch tracking the completion of all clients.
    */
-  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, int opcount, double targetperthreadperms, CountDownLatch completeLatch)
+  public ClientThread(DB db, boolean dotransactions, Workload workload, Properties props, long opcount, double targetperthreadperms, CountDownLatch completeLatch)
   {
     _db=db;
     _dotransactions=dotransactions;
@@ -400,9 +400,9 @@ class ClientThread extends Thread
   /**
    * the total amount of work this thread is still expected to do
    */
-  public int getOpsTodo()
+  public long getOpsTodo()
   {
-    int todo = _opcount - _opsdone;
+    long todo = _opcount - _opsdone;
     return todo < 0 ? 0 : todo;
   }
 }
@@ -818,7 +818,7 @@ public class Client
 
     System.err.println("Starting test.");
 
-    int opcount;
+    long opcount;
     if (dotransactions)
     {
       opcount=Integer.parseInt(props.getProperty(OPERATION_COUNT_PROPERTY,"0"));
@@ -831,7 +831,7 @@ public class Client
       }
       else
       {
-        opcount=Integer.parseInt(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
+        opcount=Long.parseLong(props.getProperty(RECORD_COUNT_PROPERTY, DEFAULT_RECORD_COUNT));
       }
     }
 
@@ -851,7 +851,7 @@ public class Client
       }
 
 
-      int threadopcount = opcount/threadcount;
+      long threadopcount = opcount/threadcount;
 
       // ensure correct number of operations, in case opcount is not a multiple of threadcount
       if (threadid<opcount%threadcount)
